[Update] Fix code review issue
