+++ static_report/src/utils/index.ts
@@ -4,4 +4,3 @@ export * from './formatters';
 export * from './store';
 export * from './theme';
 export * from './layout';
-export * from './icons';
